+++ atlasdb-cassandra-multinode-tests/src/test/java/com/palantir/cassandra/multinode/OneNodeDownTestSuite.java
@@ -15,20 +15,18 @@
  */
 package com.palantir.cassandra.multinode;
 
+import java.util.Arrays;
+
 import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.rules.RuleChain;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 import com.google.common.collect.ImmutableList;
-import com.palantir.atlasdb.containers.Containers;
 import com.palantir.atlasdb.containers.ThreeNodeCassandraCluster;
-import com.palantir.atlasdb.keyvalue.cassandra.SchemaMutationLockReleasingRule;
-import com.palantir.flake.ShouldRetry;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
+        OneNodeDownCleanCassLockStateTest.class,
         OneNodeDownGetTest.class,
         OneNodeDownPutTest.class,
         OneNodeDownMetadataTest.class,
@@ -36,20 +34,12 @@ import com.palantir.flake.ShouldRetry;
         OneNodeDownTableManipulationTest.class,
         OneNodeDownNodeAvailabilityTest.class
         })
-@ShouldRetry(numAttempts = 3)
 public final class OneNodeDownTestSuite extends NodesDownTestSetup {
 
-    // Reusing containers when the FlakeRule kicks in.
-    @ClassRule
-    public static final RuleChain ruleChain = RuleChain
-            .outerRule(new Containers(NodesDownTestSetup.class)
-                    .with(new ThreeNodeCassandraCluster()))
-            .around(SchemaMutationLockReleasingRule.createChainedReleaseAndRetry(
-                    NodesDownTestSetup::createCassandraKvs, CONFIG));
-
     @BeforeClass
-    public static void setup() {
-        NodesDownTestSetup.initializeKvsAndDegradeCluster(
+    public static void setup() throws Exception {
+        initializeKvsAndDegradeCluster(
+                Arrays.asList(OneNodeDownTestSuite.class.getAnnotation(Suite.SuiteClasses.class).value()),
                 ImmutableList.of(ThreeNodeCassandraCluster.FIRST_CASSANDRA_CONTAINER_NAME));
     }
 }