+++ atlasdb-cassandra-multinode-tests/src/test/java/com/palantir/cassandra/multinode/TwoNodesDownTestSuite.java
@@ -15,27 +15,31 @@
  */
 package com.palantir.cassandra.multinode;
 
+import java.util.Arrays;
+
 import org.junit.BeforeClass;
-import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 import com.google.common.collect.ImmutableList;
-import com.palantir.atlasdb.containers.Containers;
 import com.palantir.atlasdb.containers.ThreeNodeCassandraCluster;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses(LessThanQuorumNodeAvailabilityTest.class)
-public final class TwoNodeDownTestSuite extends NodesDownTestSetup {
-
-    @ClassRule
-    public static final Containers CONTAINERS = new Containers(NodesDownTestSetup.class)
-            .with(new ThreeNodeCassandraCluster());
+@Suite.SuiteClasses({
+        LessThanQuorumNodeAvailabilityTest.class,
+        TwoNodesDownCleanCassLockStateTest.class,
+        TwoNodesDownGetTest.class,
+        TwoNodesDownMetadataTest.class,
+        TwoNodesDownPutTest.class,
+        TwoNodesDownTableManipulationTest.class
+        })
+public final class TwoNodesDownTestSuite extends NodesDownTestSetup {
 
     @BeforeClass
-    public static void setup() {
+    public static void setup() throws Exception {
         NodesDownTestSetup.initializeKvsAndDegradeCluster(
+                Arrays.asList(TwoNodesDownTestSuite.class.getAnnotation(Suite.SuiteClasses.class).value()),
                 ImmutableList.of(ThreeNodeCassandraCluster.FIRST_CASSANDRA_CONTAINER_NAME,
-                        ThreeNodeCassandraCluster.SECOND_CASSANDRA_CONTAINER_NAME));
+                        ThreeNodeCassandraCluster.THIRD_CASSANDRA_CONTAINER_NAME));
     }
 }