+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CassandraSchemaLockCleaner.java
@@ -69,7 +69,6 @@ public final class CassandraSchemaLockCleaner {
         ExecutorService executorService = PTExecutors.newFixedThreadPool(config.poolSize(),
                 new NamedThreadFactory("Atlas CleanCassLocksState", false));
         TaskRunner taskRunner = new TaskRunner(executorService);
-        CellLoader cellLoader = new CellLoader(config, clientPool, wrappingQueryRunner, taskRunner);
 
         CellValuePutter cellValuePutter = new CellValuePutter(
                 config,
@@ -81,12 +80,19 @@ public final class CassandraSchemaLockCleaner {
 
         return new CassandraTableDropper(config,
                 clientPool,
-                cellLoader,
                 cellValuePutter,
                 wrappingQueryRunner,
                 ConsistencyLevel.ALL);
     }
 
+    /**
+     * Removes all but one lock table if multiple lock tables are present. Then releases the schema mutation lock.
+     *
+     * @throws com.palantir.common.exception.AtlasDbDependencyException if fewer than a quorum of Cassandra nodes are
+     * reachable, or the cluster cannot come to an agreement on schema versions. This method will still drop all but
+     * one lock table on each of the reachable nodes, even if there are fewer than quorum, but the schema mutation
+     * lock will not have been unlocked.
+     */
     public void cleanLocksState() throws TException {
         Set<TableReference> tables = lockTables.getAllLockTables();
         Optional<TableReference> tableToKeep = tables.stream().findFirst();