+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/keyvalue/cassandra/CellDeleter.java
@@ -40,11 +40,10 @@ import com.palantir.common.base.FunctionCheckedException;
 import com.palantir.common.base.Throwables;
 
 class CellDeleter {
+    private final CassandraClientPool clientPool;
+    private final WrappingQueryRunner wrappingQueryRunner;
+    private final ConsistencyLevel deleteConsistency;
     private final LongUnaryOperator deleteTimestampGetter;
-
-    private CassandraClientPool clientPool;
-    private WrappingQueryRunner wrappingQueryRunner;
-    private ConsistencyLevel deleteConsistency;
 
     CellDeleter(CassandraClientPool clientPool,
             WrappingQueryRunner wrappingQueryRunner,